main-circleboard 컴포넌트 분리
