+++ .vscode/settings.json
@@ -5,7 +5,7 @@
     "strings": false
   },
   "editor.suggestOnTriggerCharacters": false,
-  "editor.wordBasedSuggestions": false,
+  "editor.wordBasedSuggestions": "off",
   "editor.parameterHints.enabled": false,
   "typescript.suggest.enabled": false,
   "javascript.suggest.enabled": false
