+++ next.config.js
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  compiler: {
+    emotion: true,
+  },
+};
+
+module.exports = nextConfig;
