+++ src/app/main/hooks/useClickOverlay.ts
@@ -1,0 +1,20 @@
+import { useEffect } from "react";
+
+export const useClickOverlay = (
+  ref: React.RefObject<HTMLElement | null>,
+  handler: () => void
+) => {
+  useEffect(() => {
+    const listener = (e: MouseEvent) => {
+      if (!ref.current || ref.current.contains(e.target as Node)) return;
+      handler(); // 외부 클릭 시
+    };
+
+    document.addEventListener("mousedown", listener);
+    return () => {
+      document.removeEventListener("mousedown", listener);
+    };
+  }, [ref, handler]);
+};
+
+export default useClickOverlay;
