+++ src/app/main/page.tsx
@@ -1,5 +1,5 @@
 // /src/app/page.tsx
-import Main from "./main/Main";
+import Main from "./Main";
 
 export default function HomePage() {
   return <Main />;
