+++ src/components/Header.tsx
@@ -1,17 +1,13 @@
 "use client";
 
 import styled from "@emotion/styled";
-// import { ArrowLeft } from "lucide-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import home from "../../public/home.svg";
 import Image from "next/image";
 import { useIsMobile } from "@/hooks/useMediaQuery";
 import { useState, useEffect } from "react";
-// const StyledAnchor = styled.a`
-//   cursor: pointer;
-//   text-decoration: none;
-// `;
+
 const HeaderNav = styled.nav`
   position: fixed;
   top: 0;
